Added card prop to carousel
